Moved the leaderboard fragment files to their own subdirectory
